requirements.txt
numpy
pandas
scikit-learn
rouge-score
nltk
spacy
pyyaml
tqdm

